moved config module content to factory module
